Add files via upload
This project showcases a robust and automated solution for Linux backup and archiving using a Python script. I designed this script to streamline routine data management, ensuring both efficiency and data integrity. A key feature is the integration with Crontab, which automates the backup process and schedules the deletion of old files after 60 days, a strategic measure to optimize disk usage. This demonstrates my proficiency in Python scripting, system administration, and process automation, all of which are critical for maintaining a secure and efficient IT environment.